refactor: move files into different folders to add structure to the project .
