+++ product-service/src/entities/Product.ts
