+++ product-service/src/entities/Rating.ts
